Programatic creation of struct centric contract
allows for you to input an array of variables with only their declaration and name as structs in a txt file. The script will create an entire contract by creating internal and external get/query/update functions for each variable. The creation is modular such that you can choose how accessible the structs are dependent on the actual use case of your implementation.